blas prototype

+++ src/kernels/cpu/unary/scalar/log.c
@@ -1,7 +1,7 @@
 #include "cpu_ops.h"
 #include "macros.h"
 
-DEFINE_LOG_KERNEL(scalar_log_fp32,double)
+DEFINE_LOG_KERNEL(scalar_log_fp32,float)
 
 // void scalar_log_fp32(void* out,const void* a,double sc,uint32_t numel){
